src/test/servlets/forms/DeleteFormServletTest.java
package servlets.forms;

import org.junit.jupiter.api.Test;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;
import services.forms.DeleteFormService;

import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpServletResponse;
import java.io.IOException;
import java.sql.SQLException;

import static org.mockito.ArgumentMatchers.any;
import static org.mockito.Mockito.verify;

class DeleteFormServletTest {

    @Mock
    HttpServletResponse httpServletResponse;

    @Mock
    HttpServletRequest httpServletRequest;

    @Mock
    DeleteFormService deleteFormService;

    @InjectMocks
    DeleteFormServlet deleteFormServlet = new DeleteFormServlet();

    public void setup() throws SQLException, IOException {

        MockitoAnnotations.initMocks(this);
    }

    @Test
    void doDelete() throws SQLException, IOException {

        setup();
        //execute
        deleteFormServlet.doDelete(httpServletRequest, httpServletResponse);

        //verify
        verify(deleteFormService).execute(any(), any());
    }
}